Safari-609.1.11.1/LayoutTests/fast/dom/Geolocation/stopUpdatingForHiddenPage-expected.txt

Tests that the GeolocationClient stops updating when the page becomes hidden

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Page is initially visible
PASS document.hidden is false
PASS testRunner.isGeolocationProviderActive() is false
Adding Geolocation watcher
PASS testRunner.isGeolocationProviderActive() became true
Changing page visibility to 'hidden'
Page visibility changed to false
PASS testRunner.isGeolocationProviderActive() became false
Changing page visibility to 'visible'
Page visibility changed to true
PASS testRunner.isGeolocationProviderActive() became true
PASS successfullyParsed is true

TEST COMPLETE
